+++ base-config/asterisk/astsqlite.db.sql
@@ -563,7 +563,7 @@ CREATE TABLE groupfeatures (
  name varchar(128) NOT NULL,
  number varchar(40) NOT NULL DEFAULT '',
  context varchar(39) NOT NULL,
- timeout smallint unsigned NOT NULL DEFAULT 0,
+ timeout tinyint unsigned NOT NULL DEFAULT 0,
  deleted tinyint(1) NOT NULL DEFAULT 0,
  PRIMARY KEY(id)
 );
